Visualizationadded, joint fixed to edge
