+++ src/progeny/robotics/kinematics.py
@@ -1,5 +1,6 @@
 import typing
 
+from himeko_hypergraph.src.elements.edge import HyperEdge
 from himeko_hypergraph.src.elements.vertex import HyperVertex
 
 
@@ -10,7 +11,7 @@ class KinematicLink(HyperVertex):
         super().__init__(name, timestamp, serial, guid, suid, label, parent)
 
 
-class KinematicJoint(HyperVertex):
+class KinematicJoint(HyperEdge):
 
     def __init__(self, name: str, timestamp: int, serial: int, guid: bytes, suid: bytes, label: str,
                  parent: typing.Optional = None):
